oops, I meant float

+++ library/src/main/java/com/dd/StateManager.java
@@ -18,7 +18,7 @@ class StateManager {
         return mIsEnabled;
     }
 
-    public double getProgress() {
+    public float getProgress() {
         return mProgress;
     }
